+++ src/Sulu/Bundle/HttpCacheBundle/DependencyInjection/SuluHttpCacheExtension.php
@@ -1,0 +1,129 @@
+<?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\HttpCacheBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Loader;
+
+/**
+ * This is the class that loads and manages your bundle configuration
+ *
+ * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
+ */
+class SuluHttpCacheExtension extends Extension
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function getConfiguration(array $config, ContainerBuilder $container)
+    {
+        return new Configuration();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        $config = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
+
+        $loader = new Loader\XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader->load('event-subscribers.xml');
+        $loader->load('proxy-client.xml');
+        $loader->load('structure-cache-handlers.xml');
+
+        $this->configureProxyClient($config['proxy_client'], $container);
+        $this->configureStructureCacheHandlers($config['handlers'], $container);
+
+        $container->setAlias('sulu_http_cache.handler', 'sulu_http_cache.handler.aggregate');
+    }
+
+    /**
+     * Configure the proxy client services
+     *
+     * @param array $config
+     * @param ContainerBuilder $container
+     */
+    private function configureProxyClient($config, ContainerBuilder $container)
+    {
+        $this->configureProxyClientVarnish($config['varnish'], $container);
+
+        // get default
+        $proxyClientName = null;
+        foreach ($config as $name => $proxyClient) {
+            if ($proxyClient['enabled'] === false) {
+                continue;
+            }
+
+            if (null !== $proxyClientName) {
+                throw new InvalidConfigurationException(sprintf(
+                    'Cannot enable more than one proxy, trying to enable "%s" when "%s" is already enabled',
+                    $name,
+                    $proxyClientName
+                ));
+            }
+
+            $proxyClientName = $name;
+        }
+
+        if (null === $proxyClientName) {
+            $proxyClientName = 'symfony';
+        }
+
+        $container->setParameter('sulu_http_cache.proxy_client.name', $proxyClientName);
+        $container->setAlias('sulu_http_cache.proxy_client', 'sulu_http_cache.proxy_client.' . $proxyClientName);
+    }
+
+    /**
+     * Configure the varnish services
+     *
+     * @param array $config
+     * @param ContainerBuilder $container
+     */
+    private function configureProxyClientVarnish($config, ContainerBuilder $container)
+    {
+        foreach ($config as $key => $value) {
+            $container->setParameter(
+                $this->getAlias() . '.proxy_client.varnish.' . $key,
+                $value
+            );
+        }
+    }
+
+    /**
+     * Configure the structure cache handler services
+     *
+     * @param array $config
+     * @param ContainerBuilder $container
+     */
+    private function configureStructureCacheHandlers($config, ContainerBuilder $container)
+    {
+        $enabledHandlers = array();
+
+        // remove handlers which have not been enabled
+        foreach ($config as $handlerName => $handlerConfig) {
+            if (false === $handlerConfig['enabled']) {
+                $container->removeDefinition('sulu_http_cache.handler.' . $handlerName);
+                continue;
+            }
+
+            $enabledHandlers[] = $handlerName;
+        }
+
+        $container->setParameter('sulu_http_cache.handler.public.max_age', $config['public']['max_age']);
+        $container->setParameter('sulu_http_cache.handler.public.shared_max_age', $config['public']['shared_max_age']);
+        $container->setParameter('sulu_http_cache.handler.public.use_page_ttl', $config['public']['use_page_ttl']);
+
+        $container->setParameter('sulu_http_cache.handler.aggregate.handlers', $enabledHandlers);
+    }
+}