+++ src/Sulu/Bundle/SnippetBundle/Tests/Functional/Content/SnippetContentTest.php
@@ -137,6 +137,17 @@ class SnippetContentTest extends BaseFunctionalTestCase
         $this->assertEquals('L\'Hôtel New Hampshire', $hotel2['title']);
     }
 
+    public function testGetReferencedUuids()
+    {
+        $pageNode = $this->session->getNode('/cmf/sulu_io/contents/hotels-page');
+        $pageStructure = $this->contentMapper->loadByNode($pageNode, 'en', 'sulu_io', true, false, false);
+        $property = $pageStructure->getProperty('hotels');
+        $uuids = $this->contentType->getReferencedUuids($property);
+        $this->assertCount(2, $uuids);
+        foreach ($uuids as $uuid) {
+            $this->assertTrue(UUIDHelper::isUuid($uuid));
+        }
+    }
     public function testRemove()
     {
         $this->property->expects($this->any())