+++ src/Sulu/Component/HttpCache/Handler/PathsHandler.php
@@ -1,0 +1,105 @@
+<?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Component\HttpCache\Handler;
+
+use FOS\HttpCache\ProxyClient;
+use Sulu\Component\Content\StructureInterface;
+use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
+use Sulu\Component\HttpCache\StructureCacheHandlerInterface;
+use Symfony\Component\HttpFoundation\Response;
+use FOS\HttpCache\ProxyClient\ProxyClientInterface;
+use FOS\HttpCache\ProxyClient\Invalidation\PurgeInterface;
+use Sulu\Component\HttpCache\HandlerInvalidateStructureInterface;
+use Sulu\Component\HttpCache\HandlerFlushInterface;
+
+/**
+ * Invalidate all the paths (i.e. old and new) for a Sulu Structure
+ */
+class PathsHandler implements 
+    HandlerFlushInterface,
+    HandlerInvalidateStructureInterface
+{
+    /**
+     * @var WebspaceManagerInterface
+     */
+    private $webspaceManager;
+
+    /**
+     * @var ProxyClientInterface
+     */
+    private $proxyClient;
+
+    /**
+     * @var string
+     */
+    private $environment;
+
+    /**
+     * @var array
+     */
+    private $structuresToInvalidate = array();
+
+    /**
+     * @param WebspaceManagerInterface $webspaceManager
+     * @param ProxyClientInterface $proxyClient
+     * @param string $environment - kernel envionment, dev, prod, etc.
+     */
+    public function __construct(
+        WebspaceManagerInterface $webspaceManager,
+        PurgeInterface $proxyClient,
+        $environment
+    ) {
+        $this->webspaceManager = $webspaceManager;
+        $this->proxyClient = $proxyClient;
+        $this->environment = $environment;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function invalidateStructure(StructureInterface $structure)
+    {
+        if (!$this->proxyClient instanceof PurgeInterface) {
+            return;
+        }
+
+        $this->structuresToInvalidate[] = $structure;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function flush()
+    {
+        if (!$this->structuresToInvalidate) {
+            return;
+        }
+
+        foreach ($this->structuresToInvalidate as $structure) {
+            if (false === $structure->hasTag('sulu.rlp') ||  null === $rlp = $structure->getPropertyValueByTagName('sulu.rlp')) {
+                return;
+            }
+
+            $urls = $this->webspaceManager->findUrlsByResourceLocator(
+                $rlp,
+                $this->environment,
+                $structure->getLanguageCode(),
+                $structure->getWebspaceKey()
+            );
+
+            foreach ($urls as $url) {
+                $this->proxyClient->purge($url);
+            }
+        }
+
+        $this->proxyClient->flush();
+    }
+}