+++ tests/app/AppKernel.php
@@ -10,7 +10,6 @@ class AppKernel extends SuluTestKernel
     public function registerBundles()
     {
         $bundles = parent::registerBundles();
-        $bundles[] = new \Sulu\Bundle\SnippetBundle\SuluSnippetBundle();
         return $bundles;
     }
 }